Introduce basic concepts to the Dashboard.
Included in this commit are the ideas of a central cloud, resource
nodes, and apps.  It's a stable commit, but is precipitating a bigger
issue:

During the development of this commit the `resources` domain had to be
renamed.  The domain in question was actually `nodes` and was just
miscommunicated.  This took the opening up of several files all across
the project.

And so the decision has been reached:  The current nature-based,
Rails-inspirect folder structure is tough to maintain.  The next commit
will introduce a domain-based folder hierarchy.