Reorganization pass.
This commit restructures, well, everything.  Original story work came
pretty quickly, but later there was a need to rename a section of the
code from Resources to Nodes.  This proved pretty tough - the domain was
leaked all across the place.

It doesn't seem like organizing files by purpose (a la Rails) is a good
approach for a Redux-centric app.  So this commit moves everything into
a domain-centric organization.  Apps are in `lib/apps`, Nodes are in
`lib/nodes`, etc.

In other words, to add or remove a given domain, it should be a process
of creating the domain folder, linking up its reducer, and then
providing a default state in schema.js

In exchange, files now have their purpose tacked on the end, for
example, pages are `Name.page.js` - Where the file extension was es6, it
has now been changed to js.

Finally, most tests are disposed of.  Look for a future commit where
tests are moved to coexist in a given domain.