Revise testing architecture.
Because of the way webpack works, only those things which are explicitly
required in the development mode and deliberately built will be exposed
to a production build.

In other words, because of the new domain-specific folder architecture,
there's no longer any reason to have a completely separate spec folder
for tests.

Test files should now coexist with the build-specific files, in the same
domain folder / module.  The naming policy will be to replace the
extension like so:  {.js,-spec.js}

So, `lib/apps/apps.behavior.js` should have a spec file with this name:
`lib/apps/apps.behavior-spec.js`.

That will put the files directly next to each other in the file tree,
while simultaneously preventing the file from accidental loading in the
build process.