Visitors can create app components.
This adds another resource, scoped to Apps: components.  Right now,
behaviorally components are almost identical to Apps, except that they
have an AppId which the reducers use to power filtering functionality.

This commit also introduces a whole mess of extra tests, many of which
are the React component layer.

One major concern this commit introduces is the need to separately
document SG Components vs. React Components, especially for frontenders.
Inevitably this terminology overload is going to cause some issues - in
fact, it's already causing a heck of semantic satiation.