Components are reorderable.
Although I have the strongest notion that there's a bug underlying this
commit, I'm cutting it.

This commit includes some thorough organizational reinforcement in
preparation for a few changes and additions.

* Most links have become ButtonLinks
* Some styling has been changed to be more nested section friendly.
* Components can be dragged to reorder.
* Components on app pages are scoped to the app.

The bug I'm speaking of is a reordering bug.  It looks like the idea of
maintaining a single list of orders for components is a flawed
implementation, because multiple apps will share the same ordering list.

Here are the available approaches:

* I can redesign ordering and relationships as token members in support
  lists.
* I can investigate existing approaches such as `normalizr` and
  `reselect`.

Because I think it will save design / explanation / onboarding time down
the road, I'm going with the second option but keeping a keen eye on the
first to see if it's necessary.

An `order` token would look like this:

```javascript
{ scopeId: <SCOPE_ID>, id: <ORDER_ID>, sortedId: <SORTED_ID> }
```

The basic concept here would be to have one jumbo sorted list and then
create scopes to manage the sorting.  But it's byzantine.

And in a similar vein, a relationship token would look like this:

```javascript
{
  id: <RELATIONSHIP_ID>,
  relationships: [
    { type: 'app', id: <APP_ID> },
    { type: 'component', id: <COMPONENT_ID> }
  ]
}
```

This one is a little simpler to manage, but again, byzantine and
elaborate in exchange for a small functionality boost.  Hopefully
normalizr / reselect combos can assuage this problem.