Integrate normalizr and rework resources.
Hinted at by the previous commit, this commit implements a schema change
of a pretty large degree:

Where before resources existed in the state tree as lists, they are now
keyed maps.

This seems pretty simple, but has the added side-effect of removing so
much boilerplate; in the end, that boilerplate was probably an expression
of inexperience vs. an expression of necessity.

So, in summary, this commit does the following:

* Resource lists have been replaced with Resource maps, keyed by ID.
* Resources now have a list of all associations.
* Order is now a product of a relationship, not a meta root thing.
* So much unnecessary code is gone.
* There is a central schema, powered by
* All sagas, behavior, and reducer files have been removed.
* Relevant logic added to root-level scripts instead.
* More test coverage has been added (still need reducer tests).