Separate layout concept from page concept.
Okay, so previously the Route handler was using the Index.page component
in the apps and component folders as a layout AND a page, and the logic
was obviously suffering a bit from confusion.

Since then, I've realized the correct thing to do is to declare index
pages and use the root route component of a domain as a layout / wrapper
component.

This commit splits those two concerns up.  Hopefully this makes future
component test waves simpler.