Revise view layer.
No two ways around it - this commit is a code fist, headed straight for
the view layer.

Ok, so - a lot of previous functionality predicated the existence of
actual mocks and workflows.  So, while they may have functioned, they
weren't, specifically speaking, correct.  This commit changes that.

Introduced within:

* A visually correct menu bar
* A much trimmed routes table (there are only 3 real routes now)
* Some neat canvas particle animations

Removed:

* Tons of CRUD cruft.
* Some present, but not so valuable, tests.