Remove Modal and Keyboard controls.
What started as an effort to include external libs for modal and
keyboard support has turned into just a flat-out removal of both
features.

Basically, a timebox was hit and both features were cut for the time
being.  There is an intent to veer back around and replace the
functionality later.