Add notification display region.
Occasionally we want to create notifications or prompts for the user.
In this specific case, we want form feedback to show up whenever a form
is invalid.

This commit introduces a small notification container system, along with
the ability to add and remove notifications via Redux.  Finally, it uses
a saga to sweep every notification 3 seconds after it is displayed.