Remove all drag and drop behavior from dashboard.
Okay, Drag & Drop is pretty fantastic, and we love it - we do.  But at
this point it's introducing complexity into the overall design, we
believe, before its time.

This commit removes drag and drop behavior completely, along with all
support methods.  It also completes the work started in extracting a lot
of the core, first-attempt entity behaviors.