API integration for image repos.
Although referred to as "registries" erroneously, this commit introduces
the ability to create and destroy image repos.  A couple of things about
image repos:

* They have no fetch path currently, that will have to come later.
* They contain sensitive information (docker hub keys).
* They come on a PREBUILT api path for now, which represents a future
  feature in the core API repo.

Regarding that last point, the current path for registry/image repo
functionality is:

    /v0/registries/dockerhub/repos

But all we're really concerned with is the `repos` resource there.  The
assumption behind this design is that there is an actual OTHER resource,
called `registries`, of which `dockerhub` is a particular record of that
resource.

In the future we hope to use other resources, so this model is in play
to make that redesign more straightforward, although it does complicate
things now somewhat (for example the erroneous naming).

There is an upcoming commit post-release which should correct the
erroneous naming of registries into image repos.