Integrate releases into API.
This commit introduces releases into the API.  Entities in this specific
case are not very concerned with how a release is structured, so much as
providing entry and exit into the Redux domain via the timestamp as ID
concept.

Releases are stored using this pattern:

```
state.releases[release-app-name-release-component-name-release-timestamp]
```

It's verbose but who wants to run into the resource collision issue we
had with components again?

This commit also introduces Releases into the saga workflow in a way I
consider to be a lot cleaner.  The next commit will be reincorporating
the other sagas in the same way.  It just feels more modular.

I might also do the same with reducers, since that file is getting
pretty verbose.