Integrate instances API with logic layer.
The final (for now) API for the Dashboard is the instances API, which
provides information on actual resources which exist in a deployment.

This commit precipitates a final architectural change for this repo,
which is that most reducer behaviors will be added to a particular
subdomain for the sake of simple encapsulation.  In other words, more
of the lib subdirectory modules will have `reducer.js` files that manage
how that particular module integrates with Redux.