Create container-specific creation process.
Previously there was no tangible home for container creation in terms of
UI.  This commit creates one, and moves the container bits to their own
subfolder.

The idea I have currently about building out containers in the future is
to consider them related and with identity but not persisted.
Meanwhile, releases, which are persisted, will exist mostly as a
construct to interface with the API IO.

Shortly a volume update will come which will do something similar to
this for those records.

One troubling thing here is:  Where to store them?  If they're just
localstorage then it's per-user, which isn't very useful.  It kind of
feels like containers and volumes should be persisted somehow.  Maybe in
a container tag?  An app tag?