Creace Release dependencies: Containers, Volumes
This commit introduces Containers and Volumes to the overall dashboard
workflow.  The Dashboard's main design is to assist in the orchestration
and understanding of the data mangement involved in building a good app
release, and then after that, the dashboard should help monitor and
manage those assets.

Releases are a lengthy tree object, and built of many limbs.  For the
sake of user clarity, those limbs are centralized as Component
resources.

In the future components and volumes will be serialized into JSON
strings and stored on Component tags.  For now they are just present for
the party that created them.

Next up is a release mapper which will take a component and
intelligently build a release with its assets.