Remove Gitbook integrations.
A separate repository, `supergiant.github.io` has been stood up to be
our central organization Github page, and documentation assets have been
moved over to that repo.

This commit removes the Gitbook content from this repository in favor of
that one.

What does this mean long-term as far as documentation for this repo?

Documentation will have to be created with this repository in mind, in
specific.  There were a few earlier passes where modules of code were
over-documented, as in, highly verbose and a pain to change when the
code did.  That inline documentation has since been removed, but it will
have to be added back in a more maintainable way.

Additionally, all documentation which gets written can now focus purely
on the Dashboard as a standalone repository, which is an overall net
gain.

There may be a need to reintegrate Gitbook later, but I'd prefer to
handle that as it becomes necessary than to just assume integration from
the onset.