Add public-only dialogs to container ports.
Container ports had lagged behind the API spec somewhat - in the case of
public ports, an external entrypoint domain and port ought to have been
set.  This commit adds those features.

In the process a little refactor on the CreateContainer component has
occurred, which splits a lot of the deep object management into separate
pure functions (around 6 in total).