Establish Releases page.
Ideally this commit would have a little more to it, but en route to
composing the feature a second set of bugs has emerged which could mean
a little refactor is necessary.  Unfortunately, it's tough to debug
while focusing on releases purely.

This commit is being cut in order to clear a little cognitive load from
my skull so I can focus on that reorganization effort.  In other words,
there will be a Releases pt. 2 commit here in a little while completing
the page features.