Pages fetch resources as needed.
Step one in a redesign around eager fetching, this commit introduces
fetch resource and missing resource knowledge to several page-level
components, and also retires the concept of `page.js` style components.

In most cases there was an overly elaborate system in place where I had
originally figured we'd be using element-specific naming for components
(Planet, Star, etc.) and pure page components designed for no reason
other than rendering children components in a styled manner.

This pre-engineering obviously didn't pan out so well, so below you'll
find a ton of changes to just make things as vanilla as possible in
terms of names, in addition to a little streamlining of routes and
related components.

Finally, this commit upgrades Redux Saga.  Why?  Because there was a
race condition in the fetching system which was previously patched with
a delay script system.  The update of Redux Saga introduces `takeEvery`
watcher-style functionality, with action buffers so that an arbitrary
amount of actions can be queued up and trusted to complete reliably.