Correctly use error bodies as notification content
The error structure which the API returns looks like this:

```json
{
  "status": 500,
  "error": "Error message"
}
```

The actual error handling in use was passing this entire body to the
notification system, which was complaining (as it saw an Immutable map)
about map content.

This commit changes error handling to correctly pass the error key
content to the notification system.