Manage and view releases.
Sometimes you eat the bear, and sometimes the bear eats you.

This commit features a length of work representing a more complete
interface to Releases.  It's not entirely complete but it is usable.
Further testing will require some other features which aren't completely
made so far, and so I'm cutting it in order to get around to those.

Two things I'm not too proud of here:

* The `ReleaseDetail` component is large and heinous.  It has too many
  other components.  But, do I want to really decouple all of those into
  over a dozen *other* files, right now?  No, I don't.
* The way ReduxForm handles input normalization is a little regrettable.
  Now my reducer has to have knowledge of the forms it abstracts.  So
  the lowermost level is now officially coupled to a (potentially
  changeable) form.

Some of these things were necessary evils, but in the end I'm worn out
with fussing over them.  Hopefully I can get time for another pass soon
to iron out the kinks.