Make form fields legible and approachable for new users
This commit friendlies up a lot of forms verbiage and adds a few visual
feedback helpers.

Create Component has friendly feedback for mili-CPU that compute decimal
CPU units. RAM units are now explicitly requested in bytes, so feeback
graduates friendly numbers from decimal KB to MB to GB to TB.

Everywhere there is a heading with 0 in a list, there is now a friendly
<aside> that lets the user know what they can do with it.

Hopefully, this is all helpful stuff, and doesn't border on Clippy-level
user intervention.