Decouple submodule reducers from main reducer file
This is just a tiny incremental step towards a more testable submodule
architecture.  The problem is that there was a lot of relationship cruft
bestowed on the dashboard because of the moving target of an API spec
and the nature of the internal development cycle.

I'm imagining a more modular architecture a little more clearly now, and
one of the sticking points seems to be that aside from Client work, a
distinct module should ideally be only interested in itself and what it
needs to work.  It should offer pieces forth to other larger pages but
have no mingled dependencies.

In other words, each submodule should be a bounded context on its own.

I'd like to take a step back and do two other reinventions of the
architecture, here's the steps I'm imagining right now:

* No submodule interdependencies.
* Dashboard-specific views couple the independent pieces together with
  page-level React components.  Not sure how to handle a multiple child
  problem but I'll work it out.
* The entire client / entity layer needs to be a singular context, so
  that it can be yanked entirely out of the dashboard at a later date.

Although I don't necessarily intend to go crazy and make submodules into
their own repos just yet, it is a future possibility.  Right now that's
just about completely impossible.