Create Registry form asks for Docker credentials
In order to build a key, we're mimicking a process that other Docker Hub
tools use, wherein they ask for username and password and then base64
encode it.  This is my best guess as to how they do it based on some
source code, although it has had no real world testing as of yet.