Be tolerant of absent releases in summary.
The ReleaseSummary react component was crashing in cases where there was
neither a current nor target release for it to pull.  The cause of this
seemed to be the fact that we were attempting to call `split` on an
undefined value, expecting a string.

This commit ensures that the value being split is a string, and so nulls
in this specific instance are being tolerated a little more.