Add Travis CI config, and sutomated Docker release.
This adds Travis CI supprt for testing and Docker builds.
When a branch is pushed, Travis will automatically test the
branch, and push a container with the label unstable<branch> to
the Supergiant Dockerhub account. When master is updated, and tests
are run the latest label gets updated in Dockerhub. If a tag/release
gets set the a tag matching the release number will get pushed to
the Dockerhub.