Steamline of travis/docker build.
This change adds some build deps for npm install to the
Docker file, and reorders the travis config to place Docker
build as part of the test pass. It also places the coveralls
reporting to the post success section because the coverage api
appears to be not always 100% reliable for the coveralls nodejs
client.