Releases page should load buttons contextually.
Our release page strategy up until this point has been to just simply
show release buttons as they corresponded with the functionality we were
testing.  This is, basically, not a great or usable interface.

This commit introduces a little more feedback to the release page
buttons.  The releases page now checks to ensure that it has access to
current and target releases, and then displays relevant buttons
contextually.

In addition, the language has been a little improved to convey the idea
of releases as a queue record, not an actual entity list.