Generate correct docker keys from repo form.
Turns out the repository (nee registry) create form was only generating
one of the fields required for the docker access key.  This commit
introduces the full functionality to the mix.

One problem during the development of this feature was the discovery
that a docker key is actually whitespace significant.  In other words,
there are some ugly methods in the create registry file which speak of
this constraint.