Fix App component name displays and background star
Apps and Component detail views are supposed to show human-entered friendly
names. This commit fixes that.

Also, the places we use the  background Supergiant have increased, and in
some places, the Supergiant wasn't disappearing on view transition. A more
teardown-friendly version was needed. This commit also refactors the background
Supergiant star, so that the included library itself no longer creates an
instance of the effect. Now, it's up to the React component to create a new
instance of the class and `start()` it.