Configurable instance number for releases.
This is a hack commit introducing a configuration interactor for release
instances.  A few notes about the enclosed:

This is a PR cut in haste.  Right now there's no consideration or input
as to where the interactions should have been placed, so I cooked one up
quick and hope to have someone else actually style the interface
(ideally someone who has a clear idea of the best place to put it).

Also, it is an actual hack.  Right now there is a single entry where
instance count is stored, and it is only persisted in local storage.  So
it works for an individual experience, but needs to be reconfigured for
every component.  There's an opportunity for a quick follow-up pass
here, but it can be released now in order for a styling pass from
another team member.