Volume types correspond to AWS options.
Previous volume types were given pretty text which didn't exactly map
one to one with actual AWS options.  This commit implements that
mapping, so that created volumes don't emerge as absolute nonsense when
passed to the API.

An interesting note is that the order of the properties assigned to JSX
/ Redux Form fields is very important.  The first thing passed ought to
be the Redux Form object, and then subsequent properties will override
it.  Otherwise Redux Form will clobber your element properties.