Instance counts are keyed to app/component name.
In order to make instance counts more granular, they are now stored
based on a nesting of app and component name.

This storage approach is not my favorite, but it is a little more
expressive overall of how we think of certain tokens of metadata like
this.  I'd like to explore more restructuring of this.

However, having said that, this commit ought to be a sort of moratorium
on code features until there is test coverage and the technical debt
I've been building up can be repaid.