Support 0.5+ API with a resources js client
This commit introduces a more isolated concept of a client.  Where
before client logic and entities were interwoven with the dashboard
components and behavior, they are now being moved gradually to the
`lib/resources` folder.

The hope is that this decomposition will introduce a more modular client
which can be extracted into its own Supergiant.js repository and
probably released on NPM.

This commit does NOT integrate the client with the dashboard yet, and it
does NOT have support for app-specific resources:  containers,
components, releases, volumes, instances and deploys will have their own
commit.

The conceptual hurdle with implementing those base resources is that
they are all idiomatic to some degree.  Not to mention that they are
also child resources (or completely ephemeral), and so the additional
level of route building logic has to be considered at some point in
order to avoid a completely wrong design.