API integration, Elements folder.
Apps now use the new resources client for their functionality and have a
little more of a resilient reduce structure.  But, that isn't the
biggest change in this commit.

The biggest change here is that where possible I have done this:

Any repetitious CSS convention has been wrapped in a React component and
moved to the elements folder.

This is a huge change, and has a good degree of potential breaks which
can result from it.  But in our spike, the styles had begun to go in
every which direction.  During an audit I uncovered the following
separate strategies:

* Resource based styling (apps, components, components-create, etc.)
* Functional styling (button-group)
* Tweak and nudge styling (left-text, valign-middle)
* Core base styling (directly on elements)

Each strategy has led to its own issues, and so in the interests of
putting a stop to the bleeding before it gets worse, I believe we need
to isolate everything into one spot and do a deeper audit and decide on a
single consistent CSS strategy.