Correct extensions for two element components.
The `DeployButton` and `PlusButton` components had an erroneous capital
"C" on their extension.  In other words:

    lib/elements/DeployButton.component.js

Was actually:

    lib/elements/DeployButton.Component.js

Just about every development system this was tested on worked fine, but
the pared down container we have is not as tolerant, and it shouldn't
be.

This commit corrects those mistakes in order to allow the dashbaord
container to function correctly.