Correctly partition components on app name.
The previous method of approaching component / app sorting was a naive
approach.  It presumed that apps would have more or less distinct names.
That hasn't really been the case, however - so in instances where the
app names were very similar, a given component would appear in multiple
places.

This commit makes a little more of an intelligent guess out of the whole
thing - but ultimately what we really need to do in the long term will
be to organize components in a much more distinct way.  Either we can
give them a better keying interface on the back end or we can invent one
custom to the front-end using the tags feature.