Remove dependency on React color picker.
The latest version of React (15) had raised a peer dependency issue with
React Input Color, which relied on React 0.14 - since it's better
overall to have the current React than that particular Color Picker,
we're temporarily going to remove the Component Color feature.

A future version of the feature should be put into play which will
offer a set of select colors instead.