Correctly invoke app creation call.
In a fit of derp, a reimplementation of app creation was rolled out
which had a flawed invocation of Redux Saga's `call` effect.  Because it
was invoking a function in the flawed call, the record was still being
created - but then it was immediately throwing an error.

This commit fixes that behavior.